# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalDishes: number;
  totalCategories: number;
  inStock: number;
  outOfStock: number;
  featuredDishes: number;
  totalTables: number;
}