# lib/api-zod/src/generated/types/dish.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Dish {
  id: number;
  name: string;
  description?: string | null;
  price: number;
  isVeg: boolean;
  isAvailable: boolean;
  isFeatured: boolean;
  imageUrl?: string | null;
  categoryId: number;
  categoryName?: string | null;
}