# lib/api-zod/src/generated/types/publicSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PublicSettings {
  restaurantName: string;
  logoUrl?: string | null;
  whatsappNumber: string;
  openingHours?: string | null;
  isOpen: boolean;
  accentColor?: string | null;
}